+++ src/grepros/__main__.py
@@ -3,12 +3,12 @@
 Module entry point.
 
 ------------------------------------------------------------------------------
-This file is part of grepros - grep for ROS1 bag files and live topics.
+This file is part of grepros - grep for ROS bag files and live topics.
 Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     24.10.2021
-@modified    30.10.2021
+@modified    02.11.2021
 ------------------------------------------------------------------------------
 """
 from . import main