+++ CHANGELOG.rst
@@ -2,12 +2,13 @@
 Changelog for package grepros
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-0.1.1 (2021-11-01)
+0.1.1 (2021-11-03)
 -------------------
+* add ROS2 support
 * flush stdout on every print, to avoid buffering in redirected output
 * add --unique-only option
 * rename options --noselect-field and --noprint-field to --no-select-field and --no-print-field
 
 0.1.0 (2021-10-31)
 -------------------
-* grep for rosbag and live topics, able to print and publish and write bagfiles
+* grep for ROS1 bag files and live topics, able to print and publish and write bagfiles